linux: add header for simulation-related things
moves `getSimTimeStep` out of `host` header file
